#581975: fixed comments
