+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/src/pages/500.tsx
@@ -1,14 +1,27 @@
-import { sitecorePagePropsFactory } from 'lib/page-props-factory';
-import Layout from 'src/Layout';
-import config from 'temp/config';
-import {
-  ComponentPropsContext,
-  GraphQLErrorPagesService,
-  SitecoreContext,
-} from '@sitecore-jss/sitecore-jss-nextjs';
-import { componentFactory } from 'temp/componentFactory';
+import Head from 'next/head';
+import { GraphQLErrorPagesService, SitecoreContext } from '@sitecore-jss/sitecore-jss-nextjs';
 import { SitecorePageProps } from 'lib/page-props';
-import ServerError from 'src/ServerError';
+import Layout from 'src/Layout';
+import { componentFactory } from 'temp/componentFactory';
+import { GetStaticProps } from 'next';
+import config from 'temp/config';
+import { siteResolver } from 'lib/site-resolver';
+
+/**
+ * Rendered in case if we have 500 error
+ */
+const ServerError = (): JSX.Element => (
+  <>
+    <Head>
+      <title>500: Server Error</title>
+    </Head>
+    <div style={{ padding: 10 }}>
+      <h1>500 Internal Server Error</h1>
+      <p>There is a problem with the resource you are looking for, and it cannot be displayed.</p>
+      <a href="/">Go to the Home page</a>
+    </div>
+  </>
+);
 
 const Custom500 = (props: SitecorePageProps): JSX.Element => {
   if (!(props && props.layoutData)) {
@@ -16,31 +29,29 @@ const Custom500 = (props: SitecorePageProps): JSX.Element => {
   }
 
   return (
-    <ComponentPropsContext value={{}}>
-      <SitecoreContext componentFactory={componentFactory} layoutData={props.layoutData}>
-        <Layout layoutData={props.layoutData} headLinks={props.headLinks} />
-      </SitecoreContext>
-    </ComponentPropsContext>
+    <SitecoreContext componentFactory={componentFactory} layoutData={props.layoutData}>
+      <Layout layoutData={props.layoutData} headLinks={props.headLinks} />
+    </SitecoreContext>
   );
 };
 
-export async function getStaticProps(context: SitecorePageProps) {
-  const props = await sitecorePagePropsFactory.create(context);
+export const getStaticProps: GetStaticProps = async (context) => {
+  const site = siteResolver.getByName(config.jssAppName);
   const errorPagesService = new GraphQLErrorPagesService({
     endpoint: config.graphQLEndpoint,
     apiKey: config.sitecoreApiKey,
-    siteName: props.site.name,
-    language: (props.site.locale || context.locale) as string,
+    siteName: site.name,
+    language: context.locale || config.defaultLanguage,
   });
 
   const resultErrorPages = await errorPagesService.fetchErrorPages();
 
   return {
     props: {
-      ...props,
+      headLinks: [],
       layoutData: resultErrorPages?.serverErrorPage?.rendered || null,
     },
   };
-}
+};
 
 export default Custom500;
