--- packages/create-sitecore-jss/src/templates/nextjs/src/ServerError.tsx
@@ -1,19 +1,0 @@
-import Head from 'next/head';
-
-/**
- * Rendered in case if we have 404 error
- */
-const ServerError = (): JSX.Element => (
-  <>
-    <Head>
-      <title>500: Server Error</title>
-    </Head>
-    <div style={{ padding: 10 }}>
-      <h1>500 Internal Server Error</h1>
-      <p>There is a problem with the resource you are looking for, and it cannot be displayed.</p>
-      <a href="/">Go to the Home page</a>
-    </div>
-  </>
-);
-
-export default ServerError;
