+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/page-props.ts
@@ -1,4 +1,3 @@
-import { Redirect } from 'next';
 import {
   DictionaryPhrases,
   ComponentPropsCollection,
@@ -17,6 +16,5 @@ export type SitecorePageProps = {
   componentProps: ComponentPropsCollection;
   notFound: boolean;
   layoutData: LayoutServiceData;
-  redirect?: Redirect;
   headLinks: HTMLLink[];
 };
